RFC: Detect when suites need to be run in serial
This is an RFC style patch to run suites that define
setUpClass, tearDownClass, setUpModule and tearDownModule
in serial vis-a-vis each other.

There's a few rough edges, and a few places where I need
some comments and ideas on what to do.

The biggest missing part so far is that KeyboardInterrupt
handling isn't there. If there's a container in Python
which represents "value or exception" kind of like
multiprocessing's AsyncResult, then that'd be nice.

toParallelTestTargets subdivides a GreenTestSuite into
a set of strings representing loadable targets for the
runner. It actually subdivides all the way at the moment
until we reach something that looks like a unittest.TestCase,
at which point it checks that class and its module to see
if any serial-indicating methods are defined such that
only the class or module name should be returned.

poolRunner was also adjusted to return its results via a
queue - it may run multiple tests now and needs to indicate
to the parent runner both when a test has started and
finished.