Fix AttributeError when running non-test
The run function in runner.py assumes a specific order of messages from the queue. If the test loader got an unrunnable object, it would attempt to report it, but would never actually put the result in the queue, so the run function would fail because it assumed it would get a result instead of the sentinel None.